[FAILING] aoc1: calculate increases in list
solution for https://adventofcode.com/2021/day/1

In main.gr, change `./input` to `./example-input` to use the example
data from the problem statement. This prints `7`, which is correct.

However, using the actual input data fails with a stack overflow. Too
much recursion! This is true even if running the file with special flags
such as:

    grain --no-gc --experimental-wasm-tail-call main.gr